# server/common/swagger/Api.yaml
swagger: '2.0'
info:
  version: 1.10.1
  title: nxplorerjs-microservice
  description: Express Microservice Starter
basePath: /api/v1
...
consumes:
  - application/json
produces:
  - application/json
schemes:
  - http
  - https
securityDefinitions:
  Bearer:
    description: |
      For accessing the API a valid JWT token must be passed in all the queries in
      the 'Authorization' header.


      A valid JWT token is generated by the API and retourned as answer of a call
      to the route /login giving a valid user & password.


      The following syntax must be used in the 'Authorization' header :

          Bearer xxxxxx.yyyyyyy.zzzzzz
    type: apiKey
    name: Authorization
    in: header
definitions:
...
  ExampleBody:
    type: object
    title: example
    required:
      - name
    properties:
      name:
        type: string
        description: The example name
...
paths:
...
  /examples:
    get:
      tags:
        - Examples
      description: Fetch all examples
      security:
        - Bearer: []
      responses:
        200:
          description: Returns all examples
    post:
      tags:
        - Examples
      description: Create a new example
      security:
        - Bearer: []
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: number of items to skip
          required: true
          schema:
            $ref: '#/definitions/ExampleBody'
      responses:
        201:
          description: Example posted
        403:
          description: Forbidden Request
  /examples/{id}:
    get:
      tags:
        - Examples
      security:
        - Bearer: []
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the entity to retrieve
          type: integer
      responses:
        200:
          description: Return the example with the specified id
        404:
          description: Example not available
...